# restserver/generated/types.gen.go
// Package generated provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen DO NOT EDIT.
package generated

// Artifact defines model for Artifact.
type Artifact struct {
	Id string `json:"id"`

	// location to download the artifact using a GET request.
	Location *string `json:"location,omitempty"`
	Size     int     `json:"size"`
}

// ArtifactIds defines model for ArtifactIds.
type ArtifactIds []string

// Error defines model for Error.
type Error struct {
	Id string `json:"id"`
}

// ExtendedProject defines model for ExtendedProject.
type ExtendedProject struct {
	Description string      `json:"description"`
	Hashes      *[]Artifact `json:"hashes,omitempty"`
	Id          string      `json:"id"`
	Name        string      `json:"name"`
}

// Project defines model for Project.
type Project struct {
	Description string `json:"description"`
	Id          string `json:"id"`
	Name        string `json:"name"`
}

// ProjectCreate defines model for ProjectCreate.
type ProjectCreate struct {
	Description string `json:"description"`
	Name        string `json:"name"`
}

// Success defines model for Success.
type Success struct {
	Message string `json:"message"`
}

// CreateProjectJSONBody defines parameters for CreateProject.
type CreateProjectJSONBody ProjectCreate

// CreateProjectJSONRequestBody defines body for CreateProject for application/json ContentType.
type CreateProjectJSONRequestBody CreateProjectJSONBody